Test results review - 20
In this run, noticed quite a few failures related to nav and tab panel apps (205, 213) and also 30X bslib based apps. Most of them are cosmetic and does not seem to be the impact of any new changes. I am accepting the low impact changes and keeping the rest unaccepted to discuss with team.